% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catToDummy.R
\name{catToDummy}
\alias{catToDummy}
\title{Categorical to dummy covariates}
\usage{
catToDummy(rast)
}
\arguments{
\item{rast}{Raster of categorical covariate}
}
\value{
List of dummy covariate rasters
}
\description{
Categorical to dummy covariates
}
