% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareUD.R
\name{compareUD}
\alias{compareUD}
\title{Compare two utilisation distributions}
\usage{
compareUD(beta1, beta2, covlist)
}
\arguments{
\item{beta1}{Coefficients of the first UD}

\item{beta2}{Coefficients of the second UD}

\item{covlist}{List of covariate rasters}
}
\description{
Compare two utilisation distributions
}
