% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitRSF.R
\name{fitRSF}
\alias{fitRSF}
\title{Fit RSF}
\usage{
fitRSF(xy, method = c("poisson", "logit"), covlist, nbin = NULL,
  nzeros = NULL)
}
\arguments{
\item{xy}{Matrix of observed locations}

\item{method}{Method to fit the RSF, either "poisson" for a
Poisson GLM, or "logit" for a logistic regression}

\item{covlist}{List of covariate rasters}

\item{nbin}{Number of bins in each dimension for Poisson GLM}

\item{nzeros}{Number of zeros to sample for logistic regression}
}
\value{
Output of glm
}
\description{
Fit RSF
}
