% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMMprobs.R
\name{HMMprobs}
\alias{HMMprobs}
\title{State-dependent distributions for multistate local Gibbs}
\usage{
HMMprobs(xy, r, beta, gridc, gridz, cov, lim, res)
}
\arguments{
\item{xy}{Matrix of observed locations}

\item{r}{Vector of availability radii (one for each state)}

\item{beta}{Vector of habitat selection parameters}

\item{gridc}{Monte Carlo sample of intermediate centres c}

\item{gridz}{Monte Carlo sample of endpoints z}

\item{cov}{Array of covariates (one layer for each covariate)}

\item{lim}{Limits of the covariate rasters.}

\item{res}{Resolution of the covariate rasters.}
}
\value{
Matrix with one row for each observed step, and one column
for each state.
}
\description{
State-dependent distributions for multistate local Gibbs
}
