% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listToArray.R
\name{listToArray}
\alias{listToArray}
\title{Transform list of rasters to array}
\usage{
listToArray(covlist)
}
\arguments{
\item{covlist}{List of covariate rasters}
}
\value{
Array, for which each layer corresponds to a covariate
}
\description{
Transform list of rasters to array
}
