% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2w.R
\name{n2w}
\alias{n2w}
\title{Transform parameters from natural to working scale}
\usage{
n2w(beta, rdist = c("fixed", "exp", "gamma"), r = NULL, shape = NULL,
  rate = NULL, xy = NULL)
}
\arguments{
\item{beta}{Coefficients of the RSF}

\item{rdist}{Distribution of the availability radius ("fixed", "exp", or "gamma")}

\item{r}{Availability radius, if rdist="fixed"}

\item{shape}{Shape parameter, if rdist="gamma"}

\item{rate}{Rate parameter, if rdist="exp" or rdist="gamma"}

\item{xy}{Matrix of observed locations, needed to derive maximum step length
if rdist="fixed"}
}
\value{
Vector of parameters on the working scale
}
\description{
Transform parameters from natural to working scale
}
