% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nllkLG.R
\name{nllkLG}
\alias{nllkLG}
\title{Negative log-likelihood function for the local Gibbs model}
\usage{
nllkLG(par, ID = NULL, xy, rdist = c("fixed", "exp", "gamma", "weibull"),
  MCgrids, cov, lim, res, debug = FALSE)
}
\arguments{
\item{par}{Vector of parameters, i.e. resource selection coefficients and
parameters for the distribution of the radius.}

\item{ID}{Vector of track IDs}

\item{xy}{Matrix of observed locations}

\item{rdist}{Distribution of availability radius 
("fixed", "exp", "gamma", or "weibull")}

\item{MCgrids}{List of Monte Carlo samples, with components gridr (if needed), 
gridc, and gridz. As output by \code{\link{MCsample}}.}

\item{cov}{Array of covariates (one layer for each covariate)}

\item{lim}{Limits of the covariate rasters.}

\item{res}{Resolution of the covariate rasters.}

\item{debug}{Logical. If TRUE, the function also returns the log-likelihood
of each step}
}
\description{
Negative log-likelihood function for the local Gibbs model
}
