% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nllkLG.R
\name{nllkLG_debug}
\alias{nllkLG_debug}
\title{Negative log-likelihood function for the local Gibbs model (debug version)}
\usage{
nllkLG_debug(par, ID = NULL, xy, rdist = c("fixed", "exp", "gamma"),
  MCgrids, cov, lim, res)
}
\description{
Negative log-likelihood function for the local Gibbs model (debug version)
}
