% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nllkLG_norm_rcpp}
\alias{nllkLG_norm_rcpp}
\title{Negative log-likelihood for the local Gibbs model (C++)}
\usage{
nllkLG_norm_rcpp(beta, sigma, ID, xy, gridc, gridz, cov, lim, res)
}
\arguments{
\item{beta}{Parameters of the RSF}

\item{ID}{Vector of track IDs}

\item{xy}{Matrix of observed locations}

\item{gridc}{Grid for Monte Carlo integration}

\item{gridz}{Grid for Monte Carlo integration}

\item{cov}{Array of covariates (one layer for each covariate)}

\item{lim}{Limits of the covariate rasters.}

\item{res}{Resolution of the covariate rasters.}
}
\description{
Negative log-likelihood for the local Gibbs model (C++)
}
