% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nllkMSLG.R
\name{nllkMSLG}
\alias{nllkMSLG}
\title{Negative log-likelihood for multistate local Gibbs model}
\usage{
nllkMSLG(par, xy, nstate, MCgrids, cov, lim, res, norm = FALSE)
}
\arguments{
\item{par}{Vector of parameters on working scale}

\item{xy}{Matrix of observed locations}

\item{nstate}{Number of states}

\item{MCgrids}{List of Monte Carlo samples, with components gridc and gridz. 
As output by \code{\link{MCsample}}.}

\item{cov}{Array of covariates (one layer for each covariate)}

\item{lim}{Limits of the covariate rasters.}

\item{res}{Resolution of the covariate rasters.}

\item{norm}{Logical. TRUE if normal transition density. (Only for multistate case)}
}
\value{
Negative log-likelihood
}
\description{
Negative log-likelihood for multistate local Gibbs model
}
