% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastRSF.R
\name{rastRSF}
\alias{rastRSF}
\title{Calculate raster of RSF}
\usage{
rastRSF(beta, covlist)
}
\arguments{
\item{beta}{Coefficients of the RSF}

\item{covlist}{List of covariate rasters}
}
\value{
Raster of the RSF
}
\description{
Calculate raster of RSF
}
