% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rsfvec}
\alias{rsfvec}
\title{Resource selection function (vectorized for xy)}
\usage{
rsfvec(xy, beta, cov, lim, res)
}
\arguments{
\item{xy}{Matrix of points where the RSF should be evaluated}

\item{beta}{Vector of resource selection coefficients}

\item{cov}{Array of covariates (one layer for each covariate)}

\item{lim}{Four values: x min, x max, y min, y max}

\item{res}{Two values: resolution in x, and resolution in y}
}
\description{
Resource selection function (vectorized for xy)
}
