% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scalez}
\alias{scalez}
\title{Scale Monte Carlo sample of end points}
\usage{
scalez(gridc, gridz, r, xy0, xy1)
}
\arguments{
\item{gridc}{(Non-scaled) grid of intermediate centres, as a matrix
with two columns.}

\item{gridz}{(Non-scaled) grid of end points, as a matrix with
two columns.}

\item{r}{Availability radius}

\item{xy0}{Origin of step}

\item{xy1}{End of step}
}
\description{
Scale Monte Carlo sample of end points
}
