% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simLG_rcpp}
\alias{simLG_rcpp}
\title{Simulation function (local Gibbs sampler)}
\usage{
simLG_rcpp(nbObs, beta, allr, cov, xy0, lim, res, norm)
}
\arguments{
\item{nbObs}{Number of observations}

\item{beta}{Vector of resource selection coefficients}

\item{allr}{Vector of radii for movement kernel (of length nbObs-1)}

\item{cov}{Array of covariates (one layer for each covariate)}

\item{xy0}{Initial location}

\item{lim}{Limits of map}

\item{res}{Resolution of map}

\item{norm}{Logical (0 or 1). If TRUE, a normal transition density is used.}
}
\description{
Simulation function (local Gibbs sampler)
}
