% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSSF.R
\name{simSSF}
\alias{simSSF}
\title{Simulate from SSF}
\usage{
simSSF(nbObs, beta, xy1, xy2, nzeros, covlist, stepPar, anglePar)
}
\arguments{
\item{nbObs}{Number of observation}

\item{beta}{Vector of coefficients of the SSF}

\item{xy1}{First location}

\item{xy2}{Second location}

\item{nzeros}{Number of zeros to sample for each step}

\item{covlist}{List of covariate rasters}

\item{stepPar}{Shape and rate of gamma distribution}

\item{anglePar}{Mean and concentration of von Mises distribution}
}
\description{
Simulate from SSF
}
