% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simSSF_rcpp}
\alias{simSSF_rcpp}
\title{SSF simulation function (Fortin et al 2005)}
\usage{
simSSF_rcpp(nbObs, beta, xy1, xy2, nzeros, cov, lim, res, stepprobs, stepbreaks,
  angleprobs, anglebreaks)
}
\arguments{
\item{nbObs}{Number of observations}

\item{beta}{Vector of resource selection coefficients}

\item{xy1}{First location}

\item{xy2}{Second location}

\item{nzeros}{Number of zeros (unused steps) to sample for each step}

\item{cov}{Array of covariates (one layer for each covariate)}

\item{lim}{Limits of map}

\item{res}{Resolution of map}

\item{stepprobs}{Vector of probabilities for intervals of the range of step lengths}

\item{stepbreaks}{Vector of bounds for intervals of the range of step lengths}

\item{angleprobs}{Vector of probabilities for intervals of the range of turning angles}

\item{anglebreaks}{Vector of probabilities for intervals of the range of turning angles}
}
\value{
Matrix of simulated locations
}
\description{
SSF simulation function (Fortin et al 2005)
}
