% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simZeros_rcpp}
\alias{simZeros_rcpp}
\title{Sample zeros for SSF}
\usage{
simZeros_rcpp(nzeros, xy, stepprobs, stepbreaks, angleprobs, anglebreaks)
}
\arguments{
\item{nzeros}{Number of zeros for each observed location}

\item{xy}{Matrix of observed locations}

\item{stepprobs}{Vector of probabilities for intervals of the range of step lengths}

\item{stepbreaks}{Vector of bounds for intervals of the range of step lengths}

\item{angleprobs}{Vector of probabilities for intervals of the range of turning angles}

\item{anglebreaks}{Vector of probabilities for intervals of the range of turning angles}
}
\description{
Sample zeros for SSF
}
