% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viterbi.R
\name{viterbi}
\alias{viterbi}
\title{Viterbi algorithm for multistate local Gibbs model}
\usage{
viterbi(xy, r = NULL, sigma = NULL, beta, gamma, MCgrids, covlist,
  norm = FALSE)
}
\arguments{
\item{xy}{Matrix of observed locations}

\item{r}{Vector of availability radii (if norm=FALSE)}

\item{sigma}{Vector of standard deviations (if norm=TRUE)}

\item{beta}{Vector of selection parameters}

\item{gamma}{Transition probability matrix}

\item{MCgrids}{List of Monte Carlo samples, with components gridc 
and gridz. As output by \code{\link{MCsample}}.}

\item{covlist}{List of covariate rasters}

\item{norm}{Logical. TRUE if normal transition density. (Only for multistate case)}
}
\description{
Viterbi algorithm for multistate local Gibbs model
}
