% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCsample.R
\name{MCsample}
\alias{MCsample}
\title{Generate Monte Carlo samples}
\usage{
MCsample(nr = NULL, nc, nz, norm = FALSE)
}
\arguments{
\item{nr}{Size of Monte Carlo sample over r}

\item{nc}{Size of Monte Carlo sample over c}

\item{nz}{Size of Monte Carlo sample over z}

\item{norm}{Logical. If TRUE, standard normal bivariate
samples are returned for c and z.}
}
\value{
A list of three arrays of Monte Carlo samples:
gridr (vector), gridc (matrix), and gridz (matrix).
}
\description{
Generate Monte Carlo samples
}
