% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nllkLG_norm.R
\name{nllkLG_norm}
\alias{nllkLG_norm}
\title{Local Gibbs negative log likelihood (normal transition density)}
\usage{
nllkLG_norm(par, ID = NULL, xy, MCgrids, cov, lim, res)
}
\description{
Local Gibbs negative log likelihood (normal transition density)
}
