% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRaster.R
\name{plotRaster}
\alias{plotRaster}
\title{Plot raster with ggplot}
\usage{
plotRaster(rast, xy = NULL, norm = FALSE, log = FALSE)
}
\arguments{
\item{rast}{Raster object}

\item{xy}{Optional matrix of locations}

\item{norm}{Logical. Should the raster be normalized to integrate to 1?}

\item{log}{Logical. Should the raster be plotted on the log scale?}
}
\description{
Plot raster with ggplot
}
