% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLG.R
\name{simLG}
\alias{simLG}
\title{Generate track from local Gibbs model}
\usage{
simLG(nbObs, beta, allr, covlist, xy0 = NULL, norm = FALSE, npts = 100)
}
\arguments{
\item{nbObs}{Number of locations to simulate}

\item{beta}{Parameters of the RSF}

\item{allr}{Vector of availability radii for simulation}

\item{covlist}{List of covariate rasters}

\item{xy0}{Initial location (defaults to middle of map)}

\item{norm}{Logical. If TRUE, a normal transition density is used (with 
variance allr[1]^2).}

\item{npts}{Number of potential endpoints to sample at each time step}
}
\value{
Matrix of locations
}
\description{
Generate track from local Gibbs model
}
