% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncgridr.R
\name{truncgridr}
\alias{truncgridr}
\title{Truncate the sample of avilability radii}
\usage{
truncgridr(shape = NA, rate, ID, xy, gridr)
}
\arguments{
\item{shape}{Shape parameter of r distribution (NA if r~exp)}

\item{rate}{Rate parameter of r distribution}

\item{ID}{Vector of track IDs}

\item{xy}{Matrix of observed locations}

\item{gridr}{Grid for Monte Carlo integration}
}
\description{
Truncate the sample of avilability radii
}
