% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viterbi.R
\name{viterbi}
\alias{viterbi}
\title{Viterbi algorithm for multistate local Gibbs model}
\usage{
viterbi(xy, r, beta, gamma, MCgrids, covlist)
}
\arguments{
\item{xy}{Matrix of observed locations}

\item{r}{Vector of availability radii}

\item{beta}{Vector of selection parameters}

\item{gamma}{Transition probability matrix}

\item{MCgrids}{List of Monte Carlo samples, with components gridc 
and gridz. As output by \code{\link{MCsample}}.}

\item{covlist}{List of covariate rasters}
}
\description{
Viterbi algorithm for multistate local Gibbs model
}
